refactor(components): move Channel*.tsx to components/chat
