+++ src/components/chat/ChannelDialog.tsx
